refactor: moved header to layouts
