+++ pages/team/index.jsx
@@ -3,7 +3,7 @@ import JoinUsBanner from '../../components/Layout/JoinUsBanner';
 import Section from '../../components/Home/Section';
 import SectionContent from '../../components/Home/Section/SectionContent';
 import TeamsContent from '../../components/Teams/TeamsContent';
-import Header from '../../components/Header';
+import Header from '../../components/Layout/Header';
 
 const AboutUs = () => {
   const faculty = [
